mv server timeout func -> timeout_manager.cpp
